+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Fixed queue regression introduced in 2.0.0. It used to erase pages opportunistically for better performance.
   This was missed in a refactor. This feature is back now and unlocks a boatload of performance.
+- Fixed assert in previous release so it supports read sizes that are larger than the write size.
 
 ## 4.0.2 17-06-25